organize fs
