+++ KeyCollectorFig/KeyCollectorFig/KeyCollectorFig.vcxproj.filters
@@ -43,9 +43,6 @@
     <ClInclude Include="Operations\StartSniffHandler.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Operations\KeyboardEvent.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="Operations\StopSniffHandler.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
